Rename matmul_layernorm operator file
