+++ src/flag_gems/ops/__init__.py
@@ -245,7 +245,7 @@ from flag_gems.ops.margin_ranking_loss import margin_ranking_loss
 from flag_gems.ops.masked_fill import masked_fill, masked_fill_
 from flag_gems.ops.masked_scatter import masked_scatter, masked_scatter_
 from flag_gems.ops.masked_select import masked_select
-from flag_gems.ops.Matmul_Layer_Norm import matmul_layernorm
+from flag_gems.ops.matmul_layernorm import matmul_layernorm
 from flag_gems.ops.max import max, max_dim
 from flag_gems.ops.max_pool2d_with_indices import (
     max_pool2d_backward,
